+++ internal/server/api/apimodels/olsrd.go
@@ -18,7 +18,7 @@ type OlsrdLinkinfo struct {
 	Hostname            string  `json:"hostname"`
 	PreviousLinkStatus  string  `json:"previousLinkStatus"`
 	CurrentLinkStatus   string  `json:"currentLinkStatus"`
-	NeighborLinkQuality uint64  `json:"neighborLinkQuality"`
+	NeighborLinkQuality float32 `json:"neighborLinkQuality"`
 	SymmetryTime        uint64  `json:"symmetryTime"`
 	SeqnoValid          bool    `json:"seqnoValid"`
 	Pending             bool    `json:"pending"`
